+++ nevergrad/optimization/test_optimizerlib.py
@@ -248,3 +248,12 @@ def test_population_pickle(name: str) -> None:  # this test is added because som
     with tempfile.TemporaryDirectory() as folder:
         filepath = Path(folder) / "dump_test.pkl"
         optim.dump(filepath)
+
+
+def test_bo_instrumentation() -> None:
+    instrumentation = inst.Instrumentation(inst.var.SoftmaxCategorical([True, False]))
+    with pytest.warns(base.InefficientSettingsWarning):
+        optimizerlib.QRBO(instrumentation, budget=10)
+    with pytest.warns(None) as record:
+        optimizerlib.ParametrizedBO(gp_parameters={"alpha": 1})(instrumentation, budget=10)
+    assert not record, record.list  # no warning